+++ pkg/crd/crds/crds_test.go
@@ -31,6 +31,7 @@ import (
 	"github.com/arangodb/kube-arangodb/pkg/apis/backup"
 	"github.com/arangodb/kube-arangodb/pkg/apis/deployment"
 	"github.com/arangodb/kube-arangodb/pkg/apis/ml"
+	"github.com/arangodb/kube-arangodb/pkg/apis/platform"
 	"github.com/arangodb/kube-arangodb/pkg/apis/replication"
 	"github.com/arangodb/kube-arangodb/pkg/apis/scheduler"
 	"github.com/arangodb/kube-arangodb/pkg/apis/storage"
@@ -96,6 +97,8 @@ func Test_CRD(t *testing.T) {
 		{scheduler.DeploymentCRDName, SchedulerDeploymentDefinitionWithOptions},
 		{scheduler.BatchJobCRDName, SchedulerBatchJobDefinitionWithOptions},
 		{scheduler.CronJobCRDName, SchedulerCronJobDefinitionWithOptions},
+		{platform.ArangoPlatformStorageCRDName, PlatformStorageDefinitionWithOptions},
+		{platform.ArangoPlatformChartCRDName, PlatformChartDefinitionWithOptions},
 	}
 
 	for _, tc := range testCases {
@@ -154,6 +157,7 @@ func Test_CRDGetters(t *testing.T) {
 		AnalyticsGAEWithOptions,
 		NetworkingRouteWithOptions,
 		PlatformStorageWithOptions,
+		PlatformChartWithOptions,
 	}
 	require.Equal(t, len(AllDefinitions()), len(getters))